shorts-video-ai-quiz/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.5.0

# Image Processing
Pillow>=10.1.0

# Video Processing
moviepy>=1.0.3
numpy>=1.24.0

# Google Cloud Storage (optional - for GCS storage)
google-cloud-storage>=2.13.0

# HTTP Client (for testing)
httpx>=0.25.0
requests>=2.31.0

# SVG to PNG conversion (for emoji images)
cairosvg>=2.7.0
